+++ pkg/azuremachine/validate_create.go
@@ -10,6 +10,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	"github.com/giantswarm/azure-admission-controller/internal/errors"
+	"github.com/giantswarm/azure-admission-controller/internal/vmcapabilities"
 	"github.com/giantswarm/azure-admission-controller/pkg/generic"
 	"github.com/giantswarm/azure-admission-controller/pkg/validator"
 )
@@ -18,12 +19,14 @@ type CreateValidator struct {
 	ctrlClient client.Client
 	location   string
 	logger     micrologger.Logger
+	vmcaps     *vmcapabilities.VMSKU
 }
 
 type CreateValidatorConfig struct {
 	CtrlClient client.Client
 	Location   string
 	Logger     micrologger.Logger
+	VMcaps     *vmcapabilities.VMSKU
 }
 
 func NewCreateValidator(config CreateValidatorConfig) (*CreateValidator, error) {
@@ -36,11 +39,15 @@ func NewCreateValidator(config CreateValidatorConfig) (*CreateValidator, error) 
 	if config.Location == "" {
 		return nil, microerror.Maskf(invalidConfigError, "%T.Location must not be empty", config)
 	}
+	if config.VMcaps == nil {
+		return nil, microerror.Maskf(invalidConfigError, "%T.VMcaps must not be empty", config)
+	}
 
 	v := &CreateValidator{
 		ctrlClient: config.CtrlClient,
 		location:   config.Location,
 		logger:     config.Logger,
+		vmcaps:     config.VMcaps,
 	}
 
 	return v, nil
@@ -63,6 +70,16 @@ func (a *CreateValidator) Validate(ctx context.Context, request *v1beta1.Admissi
 	}
 
 	err = validateLocation(*cr, a.location)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	supportedAZs, err := a.vmcaps.SupportedAZs(ctx, cr.Spec.Location, cr.Spec.VMSize)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	err = validateFailureDomain(*cr, supportedAZs)
 	if err != nil {
 		return microerror.Mask(err)
 	}